Improved what was the Library menu.
Previously was difficult for user to understand how to add a module since had to right-click on background, something there were no visual hints for. So change the Library menu to 'Add Module" and added a "Add module to rack" button to make things clear and easy. Also changed names and ordering of the VCV rack items to make it clear they are for the different VCV Rack library.