Major code cleanup.
Added comments when could.

Reformatted code to follow google style.

Changed from using APP and accessing members directly. Now access goes through methods, which means that can add breakpoints or debug statements.

Changed math.hpp Vec and Rect classes so that the members are not access directly, but instead by methods.  While this change was big, it really does make the code more like proper c++

Changed use of Internal structure in all classes so that it is less accessible. Made it private and changed member name to internal_ to make it clear it is a member.

Cleaned up initialization of app.