Changed display of fps and cpu in menu bar.
Goals were to 1) display useful info; and 2) determine that info accurately and without bogging down system. This turned out to be complicated so there were a lot of changes.

For fps, changed to show actual fps and if could go to higher frame rate (which is useful info!). Also found that fps was not being determined as correctly as possible which meant that was seeing values like 29.8 fps instead of the actual 30 fps.

For cpu/meter, changed how it is determined. Now using command line "ps". This provides accurate results and doesn't bog down system.

Also made other small code improvements that encountered along the way.