Fixed updating of plugin that was broken.
For Bastl (not BastlPizza) was always getting message that plugin should be updated. But the problem was actually that the Bastl plugin library was never successfully loaded. Changed the code so that Rack doesn't tell user that broke/obsolete module should be updated when actually it never quite worked in the first place.

Also made lots of improvements to associated code to make it more understandable. Mostly improved variable names and added comments. Also changed to use getters instead of accessing variables directly.