Fixed saving of maximized window state.
For Macs at least the Window::windowMaximizeCallback() could be called twice when getting out of full screen mode: once to get out of full screen, and then a second time to make the window take up full size. This meant that the callback cannot be used to keep track of settings::windowMaximized config param.

Also, cleaned up and commented code encountered when fixing this issue.