Improved how Browser initially hidden.
Previously had to hide Browser in initUI(), which is not clear at all. But realized that can just visible_ to false for the BrowserOverlay, which is parent of Browser. Previous failure had to do with trying to do this just with Browser window.

Also cleaned up and improved some comments.